+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project (more or less) adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.3.2] - 2026-01-16
+
+### Added:
+- Option to restore default script folders
+- Confirmation dialog when deleting or restoring script folders
+
+### Fixed:
+- UI display issues when going back from ScriptLauncher
+
 ## [0.3.0] - 2026-01-14
 
 ### Added: